+++ spec/integration/shared_examples/logical_enclosure/delete.rb
@@ -1,0 +1,26 @@
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# You may not use this file except in compliance with the License.
+# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+
+RSpec.shared_examples 'LogicalEnclosureDeleteExample' do |context_name, variant|
+  include_context context_name
+
+  describe '#delete' do
+    it 'raises MethodUnavailable', if: variant == 'C7000' do
+      item = described_class.new(current_client)
+      expect { item.delete }.to raise_error(OneviewSDK::MethodUnavailable, /The method #delete is unavailable for this resource/)
+    end
+
+    it 'removes the resource', if: variant == 'Synergy' do
+      item = described_class.find_by(current_client, name: LOG_ENCL1_NAME).first
+      expect { item.delete }.not_to raise_error
+    end
+  end
+end