+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Interconnect
    - Internal Link Set
    - Logical Downlink
+   - Logical Enclosure
    - Logical Interconnect
    - Logical Interconnect Group
    - Logical Switch